fix: correct lane renderer and judge timing for long songs
- Use saturating_abs() instead of unsigned_abs() as i64 for
  bad_judge_time, preventing i64::MIN wrap-around in PMS miss POOR
  rendering.
- Change mark() parameter from i32 to i64, fixing judge scan reset
  for songs over 35 minutes where millisecond values exceed i32::MAX.
- Use milli_time() (i64) instead of time() (i32) for practice mode
  timeline display, fixing corrupted time text for songs over 35 min.
- Simplify redundant PMS stop_time calculation (algebraically
  equivalent to bad_time).

+++ crates/rubato-play/src/judge/manager/mod.rs
@@ -65,14 +65,14 @@ struct LaneIterState {
 }
 
 impl LaneIterState {
-    fn mark(&mut self, time_ms: i32, notes: &[JudgeNote]) {
+    fn mark(&mut self, time_ms: i64, notes: &[JudgeNote]) {
         while self.base_pos < self.note_indices.len().saturating_sub(1)
-            && (notes[self.note_indices[self.base_pos + 1]].time_us / 1000) < time_ms as i64
+            && (notes[self.note_indices[self.base_pos + 1]].time_us / 1000) < time_ms
         {
             self.base_pos += 1;
         }
         while self.base_pos > 0
-            && (notes[self.note_indices[self.base_pos]].time_us / 1000) > time_ms as i64
+            && (notes[self.note_indices[self.base_pos]].time_us / 1000) > time_ms
         {
             self.base_pos -= 1;
         }
